core/requirements.txt
absl-py==1.4.0
accelerate==0.18.0
addict==2.4.0
aenum==3.1.12
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
astunparse==1.6.3
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
basicsr==1.4.2
beautifulsoup4==4.12.2
blendmodes==2022
boltons==23.0.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
clean-fid==0.1.35
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
deprecation==2.1.0
dlib==19.24.1
easydict==1.10
einops==0.6.0
entrypoints==0.4
face-alignment==1.3.5
face-recognition==1.3.0
face-recognition-models==0.3.0
facexlib==0.3.0
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.11.0
filetype==1.2.0
filterpy==1.4.5
flatbuffers==23.3.3
font-roboto==0.0.1
fonts==0.0.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
gast==0.4.0
gdown==4.7.1
gfpgan==1.3.8
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==3.23.0
grpcio==1.53.0
h11==0.14.0
h5py==3.8.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.13.4
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.27.0
inflection==0.5.1
insightface==0.7.3
invisible-watermark==0.1.5
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
jsonmerge==1.9.0
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
kornia==0.6.11
lark==1.1.5
lazy_loader==0.2
libclang==16.0.0
linkify-it-py==2.0.0
llvmlite==0.39.1
lmdb==1.4.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mediapipe==0.9.2.1
ml-dtypes==0.1.0
mpmath==1.3.0
mtcnn==0.1.1
multidict==6.0.4
networkx==3.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.13.1
onnxruntime==1.14.1
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
orjson==3.8.10
packaging==23.1
pandas==2.0.0
piexif==1.1.3
Pillow==9.5.0
platformdirs==3.2.0
pooch==1.7.0
prettytable==3.7.0
protobuf==3.20.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.15.0
PyMatting==1.1.8
pyparsing==3.0.9
pyreadline3==3.4.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==1.7.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
realesrgan==0.3.0
regex==2023.3.23
rembg==2.0.32
requests==2.28.2
requests-oauthlib==1.3.1
resize-right==0.0.2
retina-face==0.0.13
rich==13.3.4
rsa==4.9
safetensors==0.3.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.26.1
sympy==1.11.1
tb-nightly==2.13.0a20230414
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.4.12
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.0
torchdiffeq==0.2.3
torchmetrics==0.11.4
torchsde==0.2.5
torchvision==0.15.1
tqdm==4.65.0
trampoline==0.1.2
transformers==4.25.1
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.1
urllib3==1.26.15
uvicorn==0.21.1
watchdog==3.0.0
wcwidth==0.2.6
websockets==11.0.1
webuiapi==0.6.2
Werkzeug==2.2.3
wrapt==1.14.1
yapf==0.32.0
yarl==1.8.2
oss2